update test structure
