+++ tests/src/test_misc.py
@@ -1,0 +1,6 @@
+import requests
+
+
+def test_languages(api):
+    x = requests.get(f"{api}/get_available_languages")
+    assert x.json() == ["English"]
